* master: (783 commits)
  Bump ora from 5.3.0 to 5.4.0 (#10223)
  Bump stylelint from 13.12.0 to 13.13.1 (#10213)
  Update SECURITY.md
  Bump apollo-server-koa from 2.23.0 to 2.24.0 (#10214)
  Bump @fingerprintjs/fingerprintjs from 3.0.7 to 3.1.1 (#10212)
  Bump react-tooltip from 4.2.17 to 4.2.18 (#10192)
  Bump eslint from 7.24.0 to 7.25.0 (#10188)
  Bump nodemailer from 6.5.0 to 6.6.0 (#10191)
  v3.6.1
  Update french translation to i18n plugin (#10152)
  Bump react-query from 3.8.3 to 3.13.12 (#10154)
  Fix #10172 (#10173)
  confirm superAdminRole exists when syncing permissions
  Bump @testing-library/jest-dom from 5.11.10 to 5.12.0 (#10166)
  Bump ioredis from 4.26.0 to 4.27.1 (#10165)
  Bump snyk from 1.551.0 to 1.566.0 (#10153)
  Bump apollo-server-koa from 2.22.1 to 2.23.0 (#10155)
  Remove cache key so the url is valid for cors policy (#10148)
  Fix issue with incorrect cli termination when using flags quickstart and no-run (#10111)
  [RBAC] Ignore only non visible attributes instead of non configurable ones (#10139)
  ...

+++ packages/strapi-connector-bookshelf/lib/queries.js
@@ -9,6 +9,9 @@ const pmap = require('p-map');
 const { convertRestQueryParams, buildQuery, escapeQuery } = require('strapi-utils');
 const { contentTypes: contentTypesUtils } = require('strapi-utils');
 const { singular } = require('pluralize');
+const { handleDatabaseError } = require('./utils/errors');
+
+const BATCH_SIZE = 1000;
 
 const { PUBLISHED_AT_ATTRIBUTE } = contentTypesUtils.constants;
 const pickCountFilters = omit(['sort', 'limit', 'start']);
@@ -50,6 +53,14 @@ module.exports = function createQueryBuilder({ model, strapi }) {
     return db.transaction(trx => fn(trx));
   };
 
+  const wrapErrors = fn => async (...args) => {
+    try {
+      return await fn(...args);
+    } catch (error) {
+      return handleDatabaseError(error);
+    }
+  };
+
   /**
    * Find one entry based on params
    */
@@ -78,12 +89,12 @@ module.exports = function createQueryBuilder({ model, strapi }) {
   /**
    * Count entries based on filters
    */
-  function count(params = {}) {
+  function count(params = {}, { transacting } = {}) {
     const filters = pickCountFilters(convertRestQueryParams(params));
 
     return model
       .query(buildQuery({ model, filters }))
-      .count()
+      .count({ transacting })
       .then(Number);
   }
 
@@ -138,7 +149,7 @@ module.exports = function createQueryBuilder({ model, strapi }) {
         return model.updateRelations({ id: entry.id, values: relations }, { transacting: trx });
       }
 
-      return this.findOne(params, null, { transacting: trx });
+      return findOne(params, null, { transacting: trx });
     };
 
     return wrapTransaction(runUpdate, { transacting });
@@ -164,7 +175,10 @@ module.exports = function createQueryBuilder({ model, strapi }) {
     return wrapTransaction(runDelete, { transacting });
   }
 
-  async function deleteMany(params, { transacting } = {}) {
+  async function deleteMany(
+    params,
+    { transacting, returning = true, batchSize = BATCH_SIZE } = {}
+  ) {
     if (params[model.primaryKey]) {
       const entries = await find({ ...params, _limit: 1 }, null, { transacting });
       if (entries.length > 0) {
@@ -173,12 +187,30 @@ module.exports = function createQueryBuilder({ model, strapi }) {
       return null;
     }
 
-    const paramsWithDefaults = _.defaults(params, { _limit: -1 });
-    const entries = await find(paramsWithDefaults, null, { transacting });
-    return pmap(entries, entry => deleteOne(entry.id, { transacting }), {
-      concurrency: 100,
-      stopOnError: true,
-    });
+    if (returning) {
+      const paramsWithDefaults = _.defaults(params, { _limit: -1 });
+      const entries = await find(paramsWithDefaults, null, { transacting });
+      return pmap(entries, entry => deleteOne(entry.id, { transacting }), {
+        concurrency: 100,
+        stopOnError: true,
+      });
+    }
+
+    // returning false, we can optimize the function
+    const batchParams = _.assign({}, params, { _limit: batchSize, _sort: 'id:ASC' });
+    // eslint-disable-next-line no-constant-condition
+    while (true) {
+      const batch = await find(batchParams, null, { transacting });
+
+      await pmap(batch, entry => deleteOne(entry.id, { transacting }), {
+        concurrency: 100,
+        stopOnError: true,
+      });
+
+      if (batch.length < BATCH_SIZE) {
+        break;
+      }
+    }
   }
 
   function search(params, populate) {
@@ -428,7 +460,6 @@ module.exports = function createQueryBuilder({ model, strapi }) {
         }
       }
     }
-    return;
   }
 
   async function deleteDynamicZoneOldComponents(entry, values, { key, joinModel, transacting }) {
@@ -461,7 +492,7 @@ module.exports = function createQueryBuilder({ model, strapi }) {
         };
       });
 
-    // verify the provided ids are realted to this entity.
+    // verify the provided ids are related to this entity.
     idsToKeep.forEach(({ id, component }) => {
       if (!allIds.find(el => el.id === id && el.component.uid === component.uid)) {
         const err = new Error(
@@ -523,7 +554,7 @@ module.exports = function createQueryBuilder({ model, strapi }) {
       .fetchAll({ transacting })
       .map(el => el.get('component_id').toString());
 
-    // verify the provided ids are realted to this entity.
+    // verify the provided ids are related to this entity.
     idsToKeep.forEach(id => {
       if (!allIds.includes(id)) {
         const err = new Error(
@@ -667,8 +698,8 @@ module.exports = function createQueryBuilder({ model, strapi }) {
   return {
     findOne,
     find,
-    create,
-    update,
+    create: wrapErrors(create),
+    update: wrapErrors(update),
     delete: deleteMany,
     count,
     search,
@@ -691,12 +722,14 @@ const buildSearchQuery = ({ model, params }) => qb => {
 
   const searchColumns = Object.keys(model._attributes)
     .filter(attribute => !associations.includes(attribute))
-    .filter(attribute => stringTypes.includes(model._attributes[attribute].type));
+    .filter(attribute => stringTypes.includes(model._attributes[attribute].type))
+    .filter(attribute => model._attributes[attribute].searchable !== false);
 
   if (!_.isNaN(_.toNumber(query))) {
     const numberColumns = Object.keys(model._attributes)
       .filter(attribute => !associations.includes(attribute))
-      .filter(attribute => numberTypes.includes(model._attributes[attribute].type));
+      .filter(attribute => numberTypes.includes(model._attributes[attribute].type))
+      .filter(attribute => model._attributes[attribute].searchable !== false);
     searchColumns.push(...numberColumns);
   }
